Update K2K documentation for usernames to have location information
The issue of usernames not showing up on horizon for federated users has
been fixed. As a result, the user does not know if they are federated
or not when using K2K (it used to show the uuid when federated). The
documentation has been updated to explain that the location info should
be added to the name so that the user knows what cloud they are currently
using. A dead link has also been updated.