Update Managing_block_storage.md
Edits for consistency and readability. Removed some of the "marketing" messaging for this to further read as a document.

Line 24 might need another revision (I'm not familiar enough with the architecture to make this edit right now, but I'd recommend removing the word "flexibility" and further clarifying the function of the scheduler for multi-backend support.